Moved factory method into PropertyModels class
Also, added comments
